MB-3094 Fix create shipment tests and move MakeAvailableMoveTaskOrder func to api_test
