+++ pkg/handlers/primeapi/api_test.go
@@ -3,6 +3,12 @@ package primeapi
 import (
 	"log"
 	"testing"
+	"time"
+
+	"github.com/gobuffalo/pop"
+
+	"github.com/transcom/mymove/pkg/models"
+	"github.com/transcom/mymove/pkg/testdatagen"
 
 	"github.com/transcom/mymove/pkg/testingsuite"
 
@@ -43,4 +49,15 @@ func TestHandlerSuite(t *testing.T) {
 
 	suite.Run(t, hs)
 	hs.PopTestSuite.TearDown()
+}
+
+// MakeAvailableMoveTaskOrder returns a default MTO with AvailableToPrimeAt set. Shared throughout this test package.
+func MakeAvailableMoveTaskOrder(db *pop.Connection) models.MoveTaskOrder {
+	now := time.Now()
+	mto := testdatagen.MakeMoveTaskOrder(db, testdatagen.Assertions{
+		MoveTaskOrder: models.MoveTaskOrder{
+			AvailableToPrimeAt: &now,
+		},
+	})
+	return mto
 }
